rename: BasicInfo 함수 이동
- 프로필 내용을 제대로 표시하지 못하는 문제를 위해 코드 리팩토링

Fixes: #37
